+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,69 +1,14 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.datasource.DataSourceUtils;
 
-import javax.sql.DataSource;
-import java.sql.Connection;
 import java.sql.SQLException;
-import java.util.NoSuchElementException;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final DataSource dataSource;
+    User findById(final long id) throws SQLException;
 
-    public UserService(
-            final UserDao userDao,
-            final UserHistoryDao userHistoryDao,
-            final DataSource dataSource
-    ) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.dataSource = dataSource;
-    }
+    void insert(final User user) throws SQLException;
 
-    public User findById(final long id) throws SQLException {
-        return userDao.findById(dataSource.getConnection(), id)
-                .orElseThrow(() -> new NoSuchElementException("해당 아이디의 사용자가 존재하지 않습니다."));
-    }
-
-    public void insert(final User user) throws SQLException {
-        final Connection conn = DataSourceUtils.getConnection(dataSource);
-        conn.setAutoCommit(false);
-        try {
-            userDao.insert(conn, user);
-
-            conn.commit();
-        } catch (SQLException e) {
-            conn.rollback();
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(conn, dataSource);
-        }
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) throws SQLException {
-        final Connection conn = DataSourceUtils.getConnection(dataSource);
-        conn.setAutoCommit(false);
-        final var user = findById(id);
-        user.changePassword(newPassword);
-        try {
-            userDao.update(conn, user);
-            userHistoryDao.log(conn, new UserHistory(user, createBy));
-
-            conn.commit();
-        } catch (SQLException | DataAccessException e) {
-            System.out.println("e = " + e);
-            conn.rollback();
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(conn, dataSource);
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createBy) throws SQLException;
 }
